requirements.txt
flask==3.1.*
gunicorn==21.*
twilio==9.*
openai-whisper==20231117
pydub==0.25.*
google-generativeai==0.8.*
langdetect==1.0.*
requests==2.32.*
python-dotenv==1.0.*
gTTS==2.5.*
